+++ syncany-messages/src/main/java/org/syncany/operations/daemon/messages/BadRequestResponse.java
@@ -1,0 +1,33 @@
+/*
+ * Syncany, www.syncany.org
+ * Copyright (C) 2011-2014 Philipp C. Heckel <philipp.heckel@gmail.com> 
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.syncany.operations.daemon.messages;
+
+import org.simpleframework.xml.Namespace;
+import org.simpleframework.xml.Root;
+
+@Root(strict = false)
+@Namespace(reference = "http://syncany.org/ws/1")
+public class BadRequestResponse extends Response {
+	public BadRequestResponse() {
+		super(400, -1, null); // Required
+	}
+	
+	public BadRequestResponse(int requestId, String message) {
+		super(400, requestId, message);
+	}
+}